Fix minor bug with complete-all-todos and delete-all-todos offline/online sync, and style app UI entirely

+++ todolist.js
@@ -1,20 +1,30 @@
 // Todolist component with props passed
 Vue.component("todolist", {
-  template: `<div class>
-    <ul class="list-group">
-      <todolist-item @remove="remove(i)" @complete="complete(i)" v-for="(item, i) in todolist" :item="item"></todolist-item>
-    </ul>
-    <button class="btn btn-primary btn-lg btn-block" v-on:click="completeAll">Complete All Todos</button>
-    <button class="btn btn-primary btn-lg btn-block" v-on:click="removeAll">Delete All Todos</button>
+  template:
+  `<div>
+    <div class="todos-list">
+      <h3 class="no-todos" v-if="todolist.length == 0">Nothing to do...?</h3>
+      <ul class="list-group list-group-flush" v-else>
+        <todolist-item @remove="remove(i)" @complete="complete(i)" v-for="(item, i) in todolist" :item="item"></todolist-item>
+      </ul>
+    </div>
+    <div class="row">
+      <div class="col-6">
+        <button class="btn btn-light" v-on:click="completeAll">Complete All &nbsp<i class="fas fa-check-square"></i></button>
+      </div>
+      <div class="col-6">
+        <button class="btn btn-light" v-on:click="removeAll">Delete All &nbsp<i class="fas fa-trash-alt"></i></button>
+      </div>
+    </div>
   <div>`,
   props: ["todolist"],
   methods: {
     complete: function(i) {
-      server.emit("complete", i);
       if (server.connected) {
         server.emit("complete", i);
       } else {
-        this.todolist[i].complete = true;
+        this.todolist[i].completed = true;
+        localStorage.todolist = JSON.stringify(this.todolist);
       }
     },
     remove: function(i) {
@@ -22,34 +32,43 @@ Vue.component("todolist", {
         server.emit("delete", i);
       } else {
         this.todolist.splice(i, 1);
+        localStorage.todolist = JSON.stringify(this.todolist);
       }
     },
     completeAll: function() {
-      server.emit("completeAll");
       if (server.connected) {
         server.emit("completeAll");
       } else {
         this.todolist.forEach(function(todo) {
           todo.completed = true;
         });
-        vm.$broadcast("persist");
+        localStorage.todolist = JSON.stringify(this.todolist);
       }
     },
     removeAll: function() {
-      server.emit("deleteAll");
+      if (server.connected) {
+        server.emit("deleteAll");
+      } else {
+        this.todolist = [];
+        localStorage.todolist = JSON.stringify(this.todolist);
+      }
     }
   }
 });
 
 // Todolist item component with props passed
 Vue.component("todolist-item", {
-  template: `<li class="list-group-item">
-  <div class="container-fluid">
-    <input @click="requestComplete" type="checkbox" v-model="item.completed">
-    <label v-if="item.completed" v-else :style="{ 'text-decoration': 'line-through' }">{{ item.title }}</label>
-    <label v-else>{{ item.title }}</label>
-    <button class="btn btn-outline-secondary" @click="requestDelete">Delete</button>
-</div>
+  template:
+  `<li class="list-group-item">
+    <div v-if="item.completed" class="todo-item">
+      <label class="completed">{{ item.title }} &nbsp<i v-bind:class="getClass(item.title)"></label>
+      <button class="btn btn-outline-secondary float-right" @click="requestDelete">Delete</button>
+    </div>
+    <div v-else>
+      <input @click="requestComplete" type="checkbox" v-model="item.completed"">
+      <label>{{ item.title }} &nbsp<i v-bind:class="getClass(item.title)"></i></label>
+      <button class="btn btn-outline-secondary float-right" @click="requestDelete">Delete</button>
+    </div>
   </li>`,
   props: ["item"],
   methods: {
@@ -58,6 +77,10 @@ Vue.component("todolist-item", {
     },
     requestDelete: function() {
       this.$emit("remove");
+    },
+    getClass(property) {
+      let icon = property.split(" ").pop().toLowerCase();
+      return "fas fa-" + icon;
     }
   }
 });
